// src/enums/higher-latitudes-adjusting.ts
'use strict';

/*
 * nemaaz
 *
 * (c) Sofiane Akbly <sofiane.akbly@digi-service.fr>
 *
 * Created by WebStorm on 27/09/2023 at 10:30
 * File /higher-latitudes-adjusting
 */

export enum HigherLatitudesAdjusting {
	NONE = 'NONE',
	MIDNIGHT = 'MIDNIGHT',
	ONE_SEVENTH = 'ONE_SEVENTH',
	ANGLE_BASED = 'ANGLE_BASED',
}